^ syntax: rust: remove block/dashed overrides, no better than stock
the anomalies during testing were because textwidth wasn't set
correctly, so it wasn't wrapping, so was fixing something that wasn't
really a problem.

for outlines, played around with this and couldn't get outlines within
comments to indent correctly no matter how it was implemented.  so the
stock is good enough for now