^ colors: smemsh256: remove so we can re-add as subdir
when we try to git subtree pull it overwrites all directories so we
can't have the other submodules sharing the directory.  instead we can
move it to a subdirectory (prefix further in) and join it there,
re-importing all the history again (we'll squash it).  that should come
in the next commit