/ rc: runtimepath: don't rely on pathogen to loud our custom colors
finding colors just means it has a $runtimepath/colors/foo.vim, so for
each of them that's being installed with bundle management, it needs to
be added to the path, in order to be found.  this also includes our own
which, although always installed -- as a merged subtree, rather than a
submodule -- is still installed in the "infect" hierarchy so it cannot
be found and thus "colorscheme 256" will not find it.

in the case that pathogen is not installed, instead we will
unconditionally load the smemsh256 color file because it will always be
there at that path.  we import it into the tree rather than
conditionally supply it via submodule only on interactive systems, the
strategy we use for other "bundles"