/ syntax: c: stop defining our own comments as it breaks formatting
breaks auto-format of comments.  unclear why we did this in the first
place, probably to try to put bulleted lists in the comment and have
them auto-format.  but the default with vim 9.1 seems to do this
anyways, did not verify but it looks like it does.

in any case it doesn't work at all without removing this, so we will
leave this commented out for now