+ rc: on start, try to to go last position in the file, if in viminfo
viminfo will save this as a register, which we jump to on BufReadPost

this makes editing a file over and over much simpler.  few downsides,
although maybe things like git commit, which always use the same
filename for commit message, might act strangely.

we'll see how this works in practice, but like it already.

this was found in vim help itself, ":help last-position-jump"