- bundle: polyglot: update from 202104 to current, which is 202210
has not itself been updated for 1.5 years, but checked a few of the
plugins it includes, and they also haven't been updated upstream.
syntax highlighting is apparently a somewhat stable enterprise... which
is sad, because python for example has changed significantly in this
timeframe, but, looking at the syntax/python.vim it has rarely changed.
this is the problem with regex-based rather than LSP-based syntax
highlighting.  unfortunately the classic vim support for LSP is weak at
the moment.  we will get there at some point.

nonetheless there have been some changes in vim-poloyglot and we are
doing a refresh at the moment so here we go