+ plugrc: polyglot: python: configure to do all highlightings
with our new scheme of files in polyglot/ dir to configure its
individual highlighting variables, we have added the
python_highlight_all variable to get every possible highlighting.  we
were using an older version of this syntax file before, ie not using the
polyglot version (for python), but rather, our own in syntax/python.vim,
and it had different variables and/or didn't have these variables.