- pack: switch bundles to vim8 package layout and remove pathogen
take advantage of the package functionality that debuted in vim8.  no
longer need vim-pathogen to add runtimes, it's done automatically by
presence in pack/foo/start/ (loads every time) and pack/foo/opt/
(requires :packadd).

this involves submodule moves.  we maintain our own .gitmodules so we
had to edit it, stash the buffer, do the "git mv" and then put our own
changes back and add them.  when doing a "git rm" followed by "git add"
these were not seen as module renames.  must be some kind of git
metadata is altered when doing the mv, presumably which would have been
fixed by a "submodule update" after commit, but we didn't want to risk
having something mess up so we let the "git mv" handle it.

there's at least one largely superfluous hierarchy level in the new
format (the directories under pack/) which are intended to group
plugins, or for people to be able to distribute "bundles of plugins,"
which are probably not necessary but we will use it to distinguish
between "bundles" that are git repos and other-packaged stuff like
vimballs (drchip distributes this way still, for example, most likely
because he is the [proud] author of the vimball format)

colorschemes can be moved into opt/ instead of start/ according to the
docs, and they are still autoloaded.  not clear why they don't need
:packadd like others in opt/.  we will probably move some of our plugins
into opt/ later that are rarely used, although they are autoload, so
it's not clear why there's even a reason for start/opt distinction
anyways.

this means there's really two levels of superfluous directories, but
that's just how upstream (ie, bram) wanted it... we could have used
symlinks to paper over this, but we'll try it for now like this and see
if we develop any other categorizations

colors were put in their own directory, like before, so we keep
them separate.  this destroys the "bundle" distinction of being
repositories, because our own repo -- smemsh256 -- is actually grafted
in with subtree split rather than as a submodule.  but we know about
this already and it's our own stuff and colors are already "special"
anyways so this is acceptable for now.

we'll run with the present layout and see how it goes...