^ rc: cmdline: don't default to cmdwin on ':' anymore, avoiding redraws
the last line on screen is already reserved for ex command line input.
once command line window was implemented (default keybind 'q:'), we
remapped command line invoke ':' to bring it up, so that it was always
used when entering a command

in practice however, the normal use case has been NOT to need any of the
command window features.  it has some speed implications because it
causes scroll/redraw to accommodate the new command line window.
especially with larger 'history' values and command line history, it
causes slowdown every time we summon the command line window.  so we
would rather just avoid it in the usual case.

mostly we wanted this for vi-keys during command line entry.  however we
normally don't need it, and for more complex scenarios, we can always
switch to the full command edit window using CTRL-F (mnemonic: "full
command window") if we need to do any command line editing, and it will
pre-fill with the current command line, even the current cursor
position is maintained.  so this just makes sense.

we'll leave the old remap commented out for now and see how it goes,
remove later if we're satisfied.

we do still want the insert-mode autocommand if we ever enter the
command line edit window (can't fathom not wanting to be in insert when
entering a command), so we keep the CmdwinEnter autocommand