+ after: ftplugin: help: make 'q' do a buffer wipeout in help buffers
we want a simple key to exit the current help.  :bw has the best
semantic that we're looking for.  :q will exit the editor if nothing has
changed, but that's not what we want.  :helpclose tries to close the
last window but that's not what we want either.  :bd leaves an unlisted
window behind.

'q' is the most obvious mapping, but there's a 'timeoutlen' delay for
that normally because of other, internally mapped commands, that we
cannot unmap to disambiguate.  so instead we reduce the timeout to 1 and
restore it to saved value, in buffer entry and leave autocommands.