+ rc: statusline: add a '*' to lhs if paste mode is active
before we disabled having ':' automatically enter command-line edit
window, we used to use that frequently just to see if paste mode was
active, because it would say "--INSERT-- (paste)" or not.

now that we would have to go to extra effort to get to full command line
window (by typing "q:" or using CTRL-F in command-line), and since we
were doing this frequently anyways, we will just add an asterisk in
status line if paste mode is active, and then we can always tell at a
glance.

proper escaping was a little tricky.  fortunately ternary operator does
not require spaces around the ? and : chars...