^ rc: viminfo: remove '%', is an issue with multiple concurrent sessions
not sure where along the way we thought it was a good idea to restore
the buffer list between invocations.  it doesn't really make sense on
its own, but then also it causes another problem: we have typically
dozens of vims open at once, all writing their buffer lists to the file,
and we end up having files in the buffer list for a particular editing
session that are totally unrelated to the current editor invocation.

so we'll just remove it, no need for this setting.