+ plugin: helpcurwin: instead of :only for help, always open in curwin
with an ':only' in the after-syntax, we had an issue sometimes, for
example in help when opening a quickfix window (for example, with
helpgrep), where vim would complain that we were "not allowed to edit
another buffer" because it ran the :only in the syntax after-file.

furthermore, we don't really want it always to be the only help window.
sometimes we have a split window and want it to be only in the
particular window we start it in.  yet :only will unconditionally use
the whole screen.

instead we try to use the tip from vim help tips.txt (see "help
help-curwin") to always open in current window.  this should be more
along the lines of what we really want.

this seems to work, but there are some unpleasant artifacts:

  - always leaves around a strange unhighlighted help.txt listed buffer
  - only one help window can be used at a time
  - goes to the right help file, but not the tag offset
  - if we were in an empty window, will go back to non-help help.txt
    buffer on help close, which is bizarre, it just sits in the file
    with no syntax highlighting inexplicably

we will try this instead and see how it works out.

unfortunately however, the user-defined command has to start with
capital letter in vim, so we cannot make it :help, have to make it
:Help (and cannot use :h).  to fix this, we also implement a so-called
"abbreviation" which will swap out :h and :help for :Help and :h for
:Help (although, it's really not because it's more like a realtime
keypress monitor-and-replace feature).