- plugin: helpcurwin: replace with better impl from stackoverflow
found someone else having this issue at stackoverflow, they came up with
a [mostly] better solution that removes the strange non-highlighting
help.txt buffer that lies around, and now goes to the right tag within
the file.  however it does not work correctly if the existing window is
less than 80 columns (it just uses the old behavior) due to a limitation
in vim (see the stackoverflow comments), and it retains the problem that
it will continue always using only one help buffer.