- plugin: helpcurwin: replace with my own impl that fixes all issues
my own implementation seems to be the simplest: just 4 lines, if you're
ok with only one help window.  if we want to support multiple windows we
need a litle more but it's straightforward.

this version seems to take care of all outstanding issues: it works with
arbitrary window widths, and you can put help in any of the windows.  so
it meets the original requirement of just replacing the current window's
buffer with the help buffer on the desired text.  all requirements
satisfied.

we have left the other two implementations in here to document the
history and we use 'finish' to exit early so those won't be parsed