- status: use simpler logic for the paste char on older vims
it will show the paste asterisk in all windows on older vims, but that's
necessary without further magic due to our use of the status variables
added in 8.1.1372.

there might be a complicated way to do it without, but we might not care
enough about those older vims.  u18 doesn't have the patch, and we use
that on oizys, so maybe we will take the time to work it out but let's
see first if we care.  even on that system we'll probably distribute our
own [statically linked] build...