- rc: colorscheme: manually include the color dir in rtp for older vims
pre-vim8, we did not have package support, so we cannot rely on
automatic addition of the color dir to runtimepath.  we used to do this
in our pathogen environment by manually 'infect'ing the color scheme
dir, but now that we rely on vim8 package support we will need to do
this manually for older vims.