+ gtags: import the template rcfile for us to modify
we will replace exuberant-ctags with universal-ctags, and then we will
change vim to go through gnu global (gtags), but back it with ctags
where global does not have a language implementation (as fallback).  and
then we can also provide overrides later when we figure out which are
better.  and try also pygments (?)

but first we need to check in the initial file so we have it in history.
it came from the distro /etc/gtags/gtags.conf