+ bundle: gutentags: add gtags management now that easytags removed
easytags+exuberant -> universal+gtags and this is the plugin for it

we will configure it in later commit.  first have to install async jobs
functionality which seems to be missing in our vim build...

gtags is supposedly better than ctags, but the way we have it set up,
gtags will fall back on universal-ctags for languages it doesn't have
its own implementation for.  and then we can continue to run cscope as a
separate tool for now, but there is also cscope integration with
gutentags which we can take a look at later also