+ plugin: gscope: new maps to use gutentags and gtags-cscope
this can use the more-powerful gnu global in combination with
universal-ctags for languages that are not known.  and the gtags-cscope.
so user can use gsX for gscope and csX keys for cscope, we'll try having
them both in parallel for a while.

todo: cscope.vim and gscope.vim are so similar, it would be good to
merge the files.  they are just strings, so they can be created
depending on an arg at call-time, or something like that