/ status: use a function to create &statusline so g:statusline_winid works
turns out the variable:

	g:statusline_winid

is only available when referencing in a function called from statusline
expansion like so: %!Function(). furthermore, the variable:

	g:actual_curwin

seems to only be available in a %{} expansion, not a %! function.  so,
we have to choose.

fortunately, comparing g:statusline_winid to win_getid() seems to work
correctly (so far).