+ status: add highlight variables so we can colorize individual elements
previously we were relying on the whole status line being StatusLine
(current window) or StatusLineNC (non-current).

now that we create it dynamically with string and variable
concatenation, we have added all distinct elements of the status line
with colorizing variables which comprise them.  we have corresponding
highlight groups and we can now tune them.