- status: do not try for partial support, latest needed vim or bust
just give a simple version of the line if we do not have all the
required functions.  rather than keep track of prerequisites and try to
do a partial implementation.  we will just test for latest version of
vim needed to build the whole thing, or else build the simpler one, and
not try to do half-assed in-between versions.

we used version check rather than check exists('g:statusline_winid')
because it will make debugging later easier when not necessary to check
for when this variable was added, against a vim version in the field