+ status: provide path for vim-provided values on older versions
for readonly, preview and filetype, we just will use the old status
%-expansions we had before this started, as alternate expansions if
fullstatus is false.  this way we always have basically the same
information even on older vim versions.  since we're using empty
highlights there also, and now with using '%' expansions, this means we
will keep a viable statusline with all StatusLine and StatusLineNC
highlight groups for the whole line.