/ status: need be relative to g:statusline_winid buffer for most things
otherwise, we are not comparing the correct window with the statusline,
because the inactive ones are not themselves current, they must be
referenced.  we have to use getbufvar() and provide the buffer, rather
than using the current-window-relative raw variable names.  this means
the name of the buffer overwrites the non-local buffers' via inactive
window statuslines