+ status: hi: move highlight defs from colorscheme to this file
this is really our 'status.vim'-specific and should not be in the
colorscheme; in this sense it's perhaps better as a plugin.

sometimes, feel maybe it does belong in the colorscheme, because that
has other highlight groups that other plugins add.

actually we will move it to plugin/ eventually, most likely.  so not
sure, but for now we will contain all colors in the same file that
defines the rest of the statusline.