- status: dynamically create all highlight strings at load time
the code is considerably shorter without triplicate-embedded tables.  we
set everything with a value in the list s:hipairs and can change
background color/attr in s:attrpairs.

we create all the strings ahead of time, then switch between which
highlight group at runtime when creating variables for the prompt string
with highlighting escapes embedded.  the non-current highlight group is
pre-calculated to rgb values 12 less, which is 2 levels darker of the
"same" color.  so current and non-current highlight differ only in
shade.

we now have succeeded defining this stuff in only one place: s:hipairs
(and s:attrpairs).  now we can modify single variable and change
everything.

this doesn't seem slow, surprisingly.  we can optimize later by having a
vim8 and vim9 version.