- status: preview: capitalize the [Preview] like the builtin %w
make it capitalized, like the built-in, for better or worse

we tried [ro] after using [RO] for some time and it looked weird.  so
maybe same will be the same for this.  although, have not seen it yet,
but we have quickfix coming up and will capitalize that as 'Fix'.