+ pack: drawit: move to different fork, later base version, more fixes
DrChip unfortunately does not use git to distribute his plugins, but
rather the "vimball" format, which is insane in 2024, but "so it
goes..."

the version we had was an older one than current (verified on DrChip's
web site), from a github-mirrored version of vim scripts scraped from
vim.org, probably before github existed.  this project was
archived/readonly and will not be scraping anymore, and vim.org scripts
are mostly dead anyways, github thankfully has replaced this mostly
stale ad-hoc collections with an easily updatable mechanism and ability
to use submodules.

however since the aforementioned vim-scripts repo is now frozen in time,
we would have to update DrChip's DrawIt plugin manually using a vimball,
and then check that into our own git.  which means it wouldn't be listed
in our .gitmodules file, which is so far exhaustive, and is used by me
as a reference to all the installed modules.

debated doing this (using the vimball), but after checking for anyone
else on github that had a mirror, found this one from user `normen` that
even contains a few other fixes and a better README, as well as stuffing
a version of the needed functions from DrChip's library 'cecutil' into
the repo, which obviates the need to have a required separate vimball or
submodule for that.

there was a one-liner fix needed which necessitated us using a fork of
our own instead though.  see normen's upstream issue 1 (from me) and the
fix number 2 (also from me, but copied from another user that had forked
normen's repo just for that fix) for explanation.