+ rc: redrawtime: increase 2s to 10s to allow for large file highlighting
we use "syntax sync fromstart" which is normally great (and vim uses
caching), but this can take a long time for really large files.  example
file is vim distro autoload/netrw.vim, use normal-mode 'G' command after
entering the file, and wait about 7 seconds (!).  10 should be enough
that it will unfreeze before trying to kill it after a presumptive hang
(if we didn't suspect syntax highlighting was taking a while)