+ plugin: helpcurwin: helpgrep: always use case-insensitive search
it's truly bizarre to need to put \c on the line in order to be case
insensitive.  "helpgrep foo\c" is totally strange and nobody would know
that and myself, will never remember this.  in any case, almost always
would most likely want case-insensitive for a grep feature where we
didn't know what we were looking for, so we'll make this the default.

conveniently we already have this helpcurwin stuff which defines a
helpgrep intermediary function... hope it's not hard to find this file
later because helpcurwin sure isn't obvious for defining
case-insensitivity for :helpgrep, but it doesn't make sense to put this
anywhere else so we will overload this file a bit for now