- plugin: gscope: replace cscope completely rather than alongside
cscope development is totally dead, whereas gnu global is developed,
albeit slowly (and with cvs, no less).

also, cscope doesn't handle spaces in filenames.  looked into fixing it,
but no way to properly do it without changing the database format.  it
currently uses spaces to delimit fields, but has a variable number of
fields at the end after the filename, so there is no good way to tell
args from filename components, a different delimiter has to be used, or
field length, or something.

in theory, this should be a complete functional replacement, with no
regression.  a lot of forum posts claim the database generation of
gscope is even better than cscope...