+ pack: bundle: bufferize: new plugin opens vim output in window
interface for :messages is horrible, something like ":vedit messages" is
needed for examining trace messages for some plugins.  there are
probably other commands that will benefit but this will be immediately
useful when examining gutentags trace output.  the plugin creating a
window with the messages (or any vim command) output