^ rc: buffertabs: remove this orphaned rcfile, plugin no longer exists
not sure where 'buffertabs' got removed, but currently there is no
occurrence of the string 'buffertabs' anywhere in the src/.vim/ tree.
it looks like it can put a buffer list in status line, but we prefer a
solution like bufexplorer for that.  so, no need for this, probably not
in the future either.  remove.