- bundle: csapprox: move to opt/ and verify still latest upstream
we probably won't use this much anymore since we have our own
colorscheme that we maintain.  no need to always load this.  also
checked upstream and it has not changed

since we will only be used with :packadd now, we removed the rcfile that
disables operation.  no need for this as we won't normally be loaded