- pack: colors: move repos back under root, with pack/colors/opt symlink
this is more like we had it before when we used pathogen.  more
straightforward: all external repos directly under colors/ and bundle/.

having a start/ and opt/ is useful, because some things are rarely used,
and could be considered optional and never loaded.  it's good to have a
place for these and a hint.  fortunately, we can still use symlinks for
this, seems to work fine