- pack: use short bundle names because they could be used with :packadd
especially for the opt/ ones, we need something we can quickly :packadd.
fortunately we have this indirection, in the form of the symlink, which
we can name anything we want, and point to the git repo (bundle) with
whatever its native case is.  this means we can keep the git repos and
.gitmodules files with their true case, while maintaining our
"abbreviations" (for :packadd) simply as symlinks in pack/*/opt/foo

the ones in pack/start/ are not as important since they will never be
:packadd'd, they are always added at startup with no need to type it