/ bundle: stabs: include my patch to fix end vs middle of line behavior
the <CR> keypress needs to behave differently if at the end of the line
and we're pressing enter.  in that case we should stand out of the way
and let the automatic indent work correctly.  however in the middle of
the line, we want to reposition at the first nonblank character and
start insert there