+ plugin: gscope: run gutentags#rescan() on every cwd change
this is helpful to connect to new databases when we switch projects, or
move around in an empty window before finding the project where we want
to start.  have to be able to run :csg command and that won't work until
rescan happens, the plugin will have old project_roots

this is equivalent of :csr abbreviation, now we don't have to enter it
manually.  we'll see if it causes any problems to have this done on cwd
change