^ plugin: funcs: don't try to alias as qfuncs, one invocation is fine
when we had originally implemented, we thought Funcs was too general,
but nonetheless it was more usable, so we thought we'll put both QFuncs
(Quickfix-rendered Functions list) and Funcs and see which one we will
keep.  now we have kept Funcs and will no longer use Qfuncs or its
abbreviation.