/ helpcurwin: fix new help within same file by re-using existing help
if we're already in a help window, we want to change what's in the
window and activate the full help logic, rather than create a new
window.  otherwise, vim doesn't change the cursor position for some
reason, it thinks it's already in the right place and so doing ":help
something" within a file where 'something' exists elsewhere within the
same file, will be essentially a no-up and never move the cursor.

it would be useful to figure out why it's doing this later, TODO