+ gtags: restore pygments, reconfigure langs, now have python gscope
pygments was hanging because it was (and afaict must) use ctags as its
backend, but had a language map that didn't correspond to ctags' own
map.  fortunately we can query ctags --list-languages and --list-maps,
but unfortunately, while Pygments has "pygmentize -L" it isn't very
helpful.  instead we manually tested all file extensions and one by one
removed them if they weren't supported, then made a combined list.

now it works on elinks source tree and linux kernel source tree

bonus, python now works with gscope.  unfortunately, golang still does
not, we will probably have to use a language server for that...

we kept the old list embedded in the file, not sure what we will do with
this in the future but good to have it as a reference.  also kept some
comments about how the work was done.