+ plugin: cmdpersist: add script to persist cmd win history
editing the command history normally only maintains the changes for the
duration of the window.  with this, both the session, and the .viminfo
file are updated, so it remains persistent.

we'll see how well this works in practice.  it's snarfed from a
stackoverflow answer and appears somewhat impromptu.  gave it a
once-over and looks ok but even author notes it will not do certain
things (he thinks?) like control characters.