^ rc: showmatch: remove as it makes screen blink on bracket inserts
dubious usefulness, and it seems to constantly make the screen blink (ie
bell) when typing in insert mode and entering, for example a right
bracket (which is not matched so it can't jump anywhere).  the docs do
actually say that it does this.

unclear why we ever wanted this option set in the first place...