^ rc: nocompatible: redundant because vim sets this if vimrc present
actually this was causing bugs because it's a composite option that was
being set in the middle of the file and was overriding some values we
had set above it.  for example, mysteriously showcmd was off but we had
set it on in the config file.  turns out it gets set off by "set
nocompatible" and that came after "set showcmd" but now it will be
set beore sourcing any files, and we'll remove this redundancy in the
process.

debated just putting it at the top, but decided against it; the behavior
of setting if vimrc file is present is straightforward from the docs and
is likely an old behavior (although we didn't check when it was added
but seems like a safe assumption).