/ rc: nocompatible: set for vim.nox which does not follow rules
the documented rules state that nocompatible is the default if a .vimrc
exists, but that isn't the case for vim.nox apparently.  doesn't really
make sense, but nonetheless it seems to work that way.  incus looks up
$VISUAL in its own path, NOT the user path, and finds vim.nox on [some
of] our systems rather than /usr/local/bin/vim, so it causes this issue
and we tracked it down to this setting.  makes no sense because .vimrc
does exist.