/ bundle: polyglot: switch to my own fork with pr878 double-load fix
since vim-polyglot is unmtaintained now, we have to cherry-pick fixes
ourselves, sadly.  this is really the first one so we'll switch to our
own which has cherry-picked the patch from upstream PR author
(https://github.com/Kache)

the purpose of the fix is to prevent double-loading of the ftdetect
fallback, which can result in some issues and performance penalties.